requirements.txt
# pandas>=1.4,<1.5
pyyaml
# jupyter
feast >= 0.20.*
faust-streaming
strawberry-graphql[debug-server]
# TikTokApi  # TikTok
tweepy # Twitter
pytest # dev
black # dev
flake8 # dev